requirements.txt
paho-mqtt==1.6.1
pkg_resources==0.0.0
